% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histo-function.R
\name{histo_function}
\alias{histo_function}
\title{Histogram Plotting Function}
\usage{
histo_function(
  dataset,
  var_of_interest,
  binwidth,
  fill_colour = "dodgerblue3",
  outline_colour = "black"
)
}
\arguments{
\item{dataset}{Dataframe containing the values of the variable of interest. Can be any
length or size, but must contain the variable of interest column name and be in a
dataframe format. Named dataset because I wanted it to be clear that we want a complete
dataset into the function. \code{data.frame}}

\item{var_of_interest}{Specific variable pulled from the data inputted in the first
parameter. This parameter should be a list of values tied with the variable of interest.
This can be inputted either directly through the use of data$var_of_interest notation, or
assigned before the function to it's own stored variable. Named var_of_interest, because
this is the variable of interest, but I wanted to shorten it slightly to make it more
readible. Didn't want to do var_of_int since int can sometimes be used as integer \code{list}}

\item{binwidth}{This is a numeric value that will specify what bin width you'd like the
histogram to employ. You can specify any width between 0 and the max value from the
variable of interest. Named for what it represents, binwidth, for clarity. \code{numeric}}

\item{fill_colour}{This is an aesthetic parameter where you can put any suitable colour
with ggplot2 in and change the appearance of the histogram fill category. The default is
\code{"dodgerblue3"}, resulting in a blue fill colour. Named since this colour will be the
one used within the ggplot2 \code{fill} call. \code{string}}

\item{outline_colour}{This is an aesthetic parameter where you can put any suitable colour
with ggplot2 in and change the appearance of the histogram outline around each bin. The
default is \code{"black"}, resulting in the a black outline. Named since this colour
will be the one used within the ggplot2 colour call to get an outline around the bins
\code{string}}
}
\value{
This function returns a histogram disaplying the distribution of the variable of
interest
}
\description{
This function takes in a dataset and a variable of interest to plot a
histogram based on the user specified bin width. The variable must be numeric, but can
have any range.
}
\examples{
set.seed(333)
random_data <- data.frame(id = 1:132, values = rnorm(n=132, mean = 23, sd = 7))

histo_function(random_data, values, 3)

histo_function(random_data, values, 5, fill_colour = "hotpink")
}
\author{
Kelsey McGuire \email{kmcgu@student.ubc.ca} \cr
}
